Add files via upload
This C++ file contains implementations of several essential binary tree operations and algorithms. It includes functions to construct a binary tree from preorder and inorder traversals, find the lowest common ancestor (LCA) of two nodes, and display the top view as well as the Kth level of the tree. Additionally, it features a method to transform the tree into a sum tree and another to generate all root-to-leaf paths. 
       The file also implements techniques to calculate the maximum width of a complete binary tree, perform Morris inorder traversal, and flatten the binary tree into a linked list. These implementations collectively demonstrate a solid grasp of binary tree structures and traversal methods.